Added LineDrawModule to manage drawing lines
The LineDrawModule controls drawing 2D lines and hooks into the
EditManager of Atlas. Currently the module is a copy of the existing
DrawModule which was created with the intent of drawing Polygons. These
two modules should be refactored into an abstract DrawModule, as well as
a PolygonDrawnModule and a LineDrawModule.